[V1] Fix int32 overflow in _num_nans logits kernel
Promote req_idx to int64 before the `req_idx * logits_stride` offset multiply,
matching the sibling sampler kernels (gumbel.py, min_p.py). Without this the
last row of a >2**31-element logits tensor is read out of bounds.

+++ vllm/v1/worker/gpu/metrics/logits.py
@@ -14,7 +14,7 @@ def _num_nans_kernel(
     vocab_size,
     BLOCK_SIZE: tl.constexpr,
 ):
-    req_idx = tl.program_id(0)
+    req_idx = tl.program_id(0).to(tl.int64)
     num_nans = 0
     for i in range(0, vocab_size, BLOCK_SIZE):
         block = i + tl.arange(0, BLOCK_SIZE)
